Team  Formation - Advanced
For an upcoming programming contest, Roy is forming some teams from the  students of his university. A team can have any number of contestants.

Roy knows the skill level of each contestant. To make the teams work as a unit, he should ensure that there is no skill gap between the contestants of the same team. In other words, if the skill level of a contestant is , then he has either the lowest skill level in his team or there exists another contestant with skill level of  in the same team. Also, no two contestants of the same team should have same skill level. Note that a contestant can write buggy code and thus can have a negative skill level.

The more contestants on the team, the more problems they can attempt at a time. So, Roy wants to form teams such that the smallest team is as large as possible.

Input Format

The first line of input contains , the number of test cases.

Each case contains an integer , the number of contestants, followed by  space separated integers. The  integer denotes the skill level of  contestant. The absolute values of skill levels will not exceed .

The total number of contestants in all cases will not exceed .

Output Format

For each test case, print the size of smallest team in a separate line.

link : https://www.hackerrank.com/challenges/team-formation